refactor: remove helpers package
