+++ internal/cli/continue.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"stackit.dev/stackit/internal/actions"
-	"stackit.dev/stackit/internal/cli/helpers"
+	"stackit.dev/stackit/internal/cli/common"
 	"stackit.dev/stackit/internal/runtime"
 )
 
@@ -19,7 +19,7 @@ func newContinueCmd() *cobra.Command {
 This command will continue the rebase and resume restacking remaining branches.`,
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, _ []string) error {
-			return helpers.Run(cmd, func(ctx *runtime.Context) error {
+			return common.Run(cmd, func(ctx *runtime.Context) error {
 				return actions.ContinueAction(ctx, actions.ContinueOptions{
 					AddAll: addAll,
 				})
