+++ internal/cli/navigation/log.go
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"stackit.dev/stackit/internal/actions"
-	"stackit.dev/stackit/internal/cli/helpers"
+	"stackit.dev/stackit/internal/cli/common"
 	"stackit.dev/stackit/internal/runtime"
 )
 
@@ -61,7 +61,7 @@ func addLogFlags(cmd *cobra.Command, f *logFlags) {
 }
 
 func executeLog(cmd *cobra.Command, f *logFlags, style string) error {
-	return helpers.Run(cmd, func(ctx *runtime.Context) error {
+	return common.Run(cmd, func(ctx *runtime.Context) error {
 		eng := ctx.Engine
 
 		// Determine branch name
