+++ internal/cli/stack/submit.go
@@ -5,7 +5,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"stackit.dev/stackit/internal/actions/submit"
-	"stackit.dev/stackit/internal/cli/helpers"
+	"stackit.dev/stackit/internal/cli/common"
 	"stackit.dev/stackit/internal/config"
 	_ "stackit.dev/stackit/internal/demo" // Register demo engine factory
 	"stackit.dev/stackit/internal/runtime"
@@ -70,7 +70,7 @@ func addSubmitFlags(cmd *cobra.Command, f *submitFlags) {
 }
 
 func executeSubmit(cmd *cobra.Command, f *submitFlags) error {
-	return helpers.Run(cmd, func(ctx *runtime.Context) error {
+	return common.Run(cmd, func(ctx *runtime.Context) error {
 		// Get config values
 		cfg, _ := config.LoadConfig(ctx.RepoRoot)
 		submitFooter := cfg.SubmitFooter()
