+++ internal/cli/stack/sync.go
@@ -5,7 +5,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"stackit.dev/stackit/internal/actions/sync"
-	"stackit.dev/stackit/internal/cli/helpers"
+	"stackit.dev/stackit/internal/cli/common"
 	"stackit.dev/stackit/internal/runtime"
 )
 
@@ -25,7 +25,7 @@ Restacks all branches in your repository that can be restacked without conflicts
 If trunk cannot be fast-forwarded to match remote, overwrites trunk with the remote version.`,
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, _ []string) error {
-			return helpers.Run(cmd, func(ctx *runtime.Context) error {
+			return common.Run(cmd, func(ctx *runtime.Context) error {
 				// Run sync action
 				return sync.Action(ctx, sync.Options{
 					All:     all,
