+++ internal/cli/track.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"stackit.dev/stackit/internal/actions"
-	"stackit.dev/stackit/internal/cli/helpers"
+	"stackit.dev/stackit/internal/cli/common"
 	"stackit.dev/stackit/internal/errors"
 	"stackit.dev/stackit/internal/runtime"
 )
@@ -22,7 +22,7 @@ func newTrackCmd() *cobra.Command {
 		Long: `Start tracking the current (or provided) branch with stackit by selecting its parent.
 Can recursively track a stack of branches by specifying each branch's parent interactively.
 This command can also be used to fix corrupted stackit metadata.`,
-		ValidArgsFunction: helpers.CompleteBranches,
+		ValidArgsFunction: common.CompleteBranches,
 		SilenceUsage:      true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			// Get context (demo or real)
@@ -56,7 +56,7 @@ This command can also be used to fix corrupted stackit metadata.`,
 	cmd.Flags().BoolVarP(&force, "force", "f", false, "Sets the parent to the most recent tracked ancestor of the branch being tracked to skip prompts. Takes precedence over --parent")
 	cmd.Flags().StringVarP(&parent, "parent", "p", "", "The tracked branch's parent. Must be set to a tracked branch. If provided, only one branch can be tracked at a time.")
 
-	_ = cmd.RegisterFlagCompletionFunc("parent", helpers.CompleteBranches)
+	_ = cmd.RegisterFlagCompletionFunc("parent", common.CompleteBranches)
 
 	return cmd
 }
